backend/foodgram/requirements.txt
Django==3.2.15
pytz==2022.2.1
sqlparse==0.4.2
Pillow==9.2.0
djangorestframework==3.13.1
django-filter==22.1
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
django-cors-headers == 3.13.0